Renamed Modal. Removed unused passthroughId in Asset Metadata
