+++ apps/mux/frontend/src/components/AssetConfiguration/MetadataConfiguration.tsx
